update backend for add scheduler for send daily report/db get hour/ handler fort api userH and frontend for get hour
