+++ frontend/gnomonitoring/app/api/add-webhook/route.ts
@@ -1,3 +1,5 @@
+// app/api/add-webhooks/route.ts
+
 import { NextRequest, NextResponse } from 'next/server'
 
 export async function POST(req: NextRequest) {
