+++ frontend/gnomonitoring/app/api/create-user/route.ts
@@ -1,3 +1,5 @@
+// app/api/create-user/route.ts
+
 import { auth } from "@clerk/nextjs/server";
 import { NextResponse } from "next/server";
