+++ frontend/gnomonitoring/app/api/update-report-hour/route.ts
@@ -4,6 +4,9 @@ import { NextRequest, NextResponse } from "next/server";
 export async function PUT(req: NextRequest) {
     const body = await req.json();
     const { hour, minute, user_id } = body;
+    // Just for test 
+    const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
+    // ===============================
 
     if (hour === undefined || minute === undefined || !user_id) {
         return new NextResponse("Paramètres manquants", { status: 400 });
@@ -14,7 +17,7 @@ export async function PUT(req: NextRequest) {
         const res = await fetch(`${backendURL}/usersH`, {
             method: "PUT",
             headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ hour, minute, user_id }),
+            body: JSON.stringify({ hour, minute, timezone, user_id }),
         });
 
         const data = await res.text();
